+++ tests/profiles/adobefonts_test.py
@@ -50,6 +50,7 @@ def test_get_family_checks():
         f"com.google.fonts/check/family/win_ascent_and_descent{OVERRIDE_SUFFIX}",
         "com.google.fonts/check/family/vertical_metrics",
         "com.google.fonts/check/family/single_directory",
+        "com.adobe.fonts/check/family/consistent_family_name",
         # should it be included here? or should we have
         # a get_superfamily_checks() method?
         # 'com.google.fonts/check/superfamily/vertical_metrics',
@@ -60,7 +61,7 @@ def test_get_family_checks():
 def test_profile_check_set():
     """Confirm that the profile has the correct number of checks and the correct
     set of check IDs."""
-    assert len(SET_EXPLICIT_CHECKS) == 79
+    assert len(SET_EXPLICIT_CHECKS) == 80
     explicit_with_overrides = sorted(
         f"{check_id}{OVERRIDE_SUFFIX}" if check_id in OVERRIDDEN_CHECKS else check_id
         for check_id in SET_EXPLICIT_CHECKS